Fix: Recv, Monitor PD Data Type 수정 (uint16_t -> int16_t)

+++ Core/App_Src/driver/drv_adc_ads130b04.h
@@ -54,7 +54,7 @@ typedef enum {
 HAL_StatusTypeDef DRV_ADS130B04_Init(SPI_HandleTypeDef *p_handle, HalPdMeasRespCb_t cb_fn);
 HAL_StatusTypeDef DRV_ADS130B04_Start(void);
 HAL_StatusTypeDef DRV_ADS130B04_Stop(void);
-HAL_StatusTypeDef DRV_ADS130B04_GetData(Ads130b04ChSel_t ch, double *p_data);
+HAL_StatusTypeDef DRV_ADS130B04_GetData(Ads130b04ChSel_t ch, int16_t *p_data);
 HAL_StatusTypeDef DRV_ADS130B04_Read(void);
 
 #ifdef __cplusplus
